Refactor mail (#312)
* fixes #311

* update composer url | moved some tests

* updating tests and classes

* update tests

* adding php 7.1

* added declare(strict_types=1);

* added new message contract

* added postmark driver and tests

* cs fixes

* fixed some tests and vars

* fixed more tests

* fixed more tests

* fixed tests

* fixed test

* update tests

* remove ext-curl

* fix test

* small changes

* update composer require

* cs fixes | Removed all HHVM code

+++ src/Viserio/Http/Tests/UploadedFileTest.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 namespace Viserio\Http\Tests;
 
 use ReflectionProperty;
@@ -20,8 +21,9 @@ class UploadedFileTest extends \PHPUnit_Framework_TestCase
     public function tearDown()
     {
         foreach ($this->cleanup as $file) {
-            if (is_scalar($file) && file_exists($file)) {
+            if (is_string($file) && is_scalar($file) && file_exists($file)) {
                 unlink($file);
+                $this->cleanup = [];
             }
         }
     }
